Change return value of audioHardwareSampleRate from double to float.
Change need to compile with newer version of Nix.

Reviewed by: Marcelo Lira

+++ src/ContentsInjectedBundle/PlatformClient.h
@@ -10,7 +10,7 @@ public:
     PlatformClient();
 
     // Audio --------------------------------------------------------------
-    virtual double audioHardwareSampleRate() { return 44100; }
+    virtual float audioHardwareSampleRate() { return 44100; }
     virtual size_t audioHardwareBufferSize() { return 128; }
     virtual unsigned audioHardwareOutputChannels() { return 2; }
